More DP problems

+++ Data structures and algorithms/BGCODER/CatalanNumbers/CatalanAndMath.cs
@@ -1,22 +1,16 @@
 ﻿using System;
-using System.Linq;
-
 class CatalanAndMath
 {
     static void Main()
     {
         // In the combinatorial mathematics, the Catalan numbers are calculated by the following formula:
-
-        double twoTimesNFact = 1; // I asign 1 to those variables since factorial calculations of 1 and 0 are useless
-        double nPlusOneFact = 1;
-        double nFact = 1;
-
-        var userInputN = int.Parse(Console.ReadLine());
-
-        for (int i = 1; i <= userInputN * 2; i++) // This is used to calculate 2n!
+        ulong twoTimesNFact = 1; // I asign 1 to those variables since factorial calculations of 1 and 0 are useless
+        ulong nPlusOneFact = 1;
+        ulong nFact = 1;
+        var userInputN = ulong.Parse(Console.ReadLine());
+        for (ulong i = 1; i <= userInputN * 2; i++) // This is used to calculate 2n!
         {
             twoTimesNFact = twoTimesNFact * i;
-
             if (i <= userInputN) // This calculates n!
             {
                 nFact = nFact * i;
@@ -26,6 +20,6 @@ class CatalanAndMath
                 nPlusOneFact = nPlusOneFact * i;
             }
         }
-        Console.WriteLine((twoTimesNFact/(nPlusOneFact*nFact)));
+        Console.WriteLine((twoTimesNFact / (nPlusOneFact * nFact)));
     }
 }
